refactor Hyperlink stuff in seperate package
